s3_pvc/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-kubernetes>=4.0.0,<5.0.0
pulumi-aws>=6.0.2,<7.0.0
python-dotenv==1.0.0